Got the "ps" command to work properly.
The next step should be to work on the next two parts of this
assignment. Something a bit more concerning is that xv6 hangs due to
some trap being triggered when running the testing program p2testsuite.
So although the functionality is there, I guess it's not up to standard
or something.