Made some small fixes in exec.c and chmod.c
exec.c was changed so that it uses the stat structure rather than the
inode structure. This decreases the number of header files needed in
exec.c, and allows exec.c to have better separation from datatypes it
doesn't need to know about.

chmod.c was changed so that it reads input as octal rather than decimal.